physfs_platform_ogc.c: new backend for Wii and GameCube
Add support for the Nintendo Wii and GameCube consoles. Most of the code
has been copied from the POSIX backend, but libpthread has been replaced
with libogc's functions.

The computation of directory paths is a bit peculiar, in that the Wii
and GameCube don't have the concept of users.

The base directory is computed from the current working directory, which
is set by the libfat library. Applications not using libfat have to take
care of setting the current working directory themselves.